requirements.txt
py-cord==2.4.1
openai==0.27.4
python-dotenv==1.0.0
pynacl
pytz
datetime